+++ app/controllers/v2/AttachmentsController.scala
@@ -104,7 +104,7 @@ class AttachmentsController @Inject() (
 
   def uploadAttachment(reference: String): Action[Either[MaxSizeExceeded, MultipartFormData[TemporaryFile]]] =
     (verify.authenticated andThen verify.casePermissions(reference) andThen verify.mustHave(Permission.ADD_ATTACHMENT))
-      .async(parse.maxLength(appConfig.fileUploadMaxSize, parse.multipartFormData)) { implicit request =>
+      .async(parse.maxLength(appConfig.fileUploadMaxSize.toLong, parse.multipartFormData)) { implicit request =>
         request.body match {
           case Left(MaxSizeExceeded(_)) =>
             renderAttachmentsErrors(
